change output format to integer

+++ Algorithmic Toolbox/Algorithmic Warm Up/Least Common Multiple/lcm.py
@@ -20,7 +20,7 @@ def gcd( a, b):
 def lcm(a, b):
     assert 1 <= a <= 2 * 10 ** 9 and 1 <= b <= 2 * 10 ** 9
 
-    return (a*b) / gcd(a,b)
+    return int((a*b) / gcd(a,b))
 
 
 if __name__ == '__main__':
